Refactor task pool to handle errors and report them back to the queue, allowing all processables to throw exceptions
The TaskPool has been renamed to the ProcessPool.
The ProcessPool is able to tell which processable threw by using a dictionary to map the task back to the processable.
In the process of this refactor the ProcessPool has been made no longer generic, it doesn't need to be for this project, simplicity is key.
This now allows all processables to throw if they wish, this is great because they all had a lot of repeated error handling.
They can throw cancellation exceptions to be discarded, gone exceptions discard them too, unknown exceptions cause retries. Handling rate limiting exceptions can come soon.